error handling

+++ src/controllers/img/deleteImg.js
@@ -1,9 +1,18 @@
 const db = require('../../entities/Database');
+const { BadRequestApiError } = require('../../validators/errors/ApiError');
 
-module.exports = async (req, res) => {
-  const imgId = req.params.id;
+module.exports = async (req, res, next) => {
+  try {
+    if (!req.params.id) {
+      throw new BadRequestApiError('Image id should be provided.');
+    }
 
-  const id = await db.remove(imgId);
+    const imgId = req.params.id;
 
-  return res.json({ id });
+    const id = await db.remove(imgId);
+
+    return res.json({ id });
+  } catch (err) {
+    next(err);
+  }
 };
